requirements.txt
streamlit
pandas
gspread
google-auths